build(backend,server): compile for narrow-CK_ULONG targets
ADR-0011: the backend FFI layer and server must build on narrow-CK_ULONG
targets too, both to run the i686 integration tests and for the reverse
direction (64c/32b -- a 32-bit or Windows LLP64 server proxying a backend whose
CK_ULONG is 32-bit). Widen the u64 wire-newtype constructions and CK_RV
comparisons (compare in native CK_RV width via `as cryptoki_sys::CK_RV`), and
allow clippy::unnecessary_cast + clippy::useless_conversion crate-wide.

The full workspace now builds for i686-unknown-linux-gnu (0 errors); x86_64
clippy -D warnings stays clean. The Windows server OS port (cfg-gating the
Unix-only listener/peer-cred paths) is tracked separately.

+++ crates/backend/src/ffi.rs
@@ -1431,7 +1431,7 @@ mod tests {
             backend
                 .mech_cache
                 .insert(session, ffi_conversion::mechanism_to_ffi(&mechanism).unwrap());
-            backend.remember_session_slot(CkSessionHandle(session), CkSlotId(slot));
+            backend.remember_session_slot(CkSessionHandle(session as u64), CkSlotId(slot as u64));
         }
 
         backend.drop_mech_cache_for_slot(CkSlotId(11));
